Enable more models' benchmarking for BF16

+++ torchbenchmark/models/resnext50_32x4d/__init__.py
@@ -4,6 +4,6 @@ from torchbenchmark.tasks import COMPUTER_VISION
 class Model(TorchVisionModel):
     task = COMPUTER_VISION.CLASSIFICATION
 
-    def __init__(self, device=None, jit=False, fuser="", train_bs=8, eval_bs=8, extra_args=[]):
-        super().__init__(model_name="resnext50_32x4d", device=device, jit=jit, fuser=fuser,
+    def __init__(self, device=None, jit=False, fuser="", dtype="fp32", train_bs=8, eval_bs=8, extra_args=[]):
+        super().__init__(model_name="resnext50_32x4d", device=device, jit=jit, fuser=fuser, dtype=dtype,
                          train_bs=train_bs, eval_bs=eval_bs, extra_args=extra_args)
